Add file icon for jxl files
It'd be great if it were possible to directly reference the "image" icon
from the Seti theme, but it doesn't look like that is possible.
Adding to/extending the Seti theme is also not possible. Defining a new
file icon theme
(https://code.visualstudio.com/api/extension-guides/file-icon-theme)
also doesn't work as this would require me to redefine every existing
file icon, and make users select the JPEG XL theme instead of their
usual one.

The only workable solution seems to be to define a custom language that
references a copy of
https://github.com/jesseweed/seti-ui/blob/master/icons/image.svg, albeit
with spacing adjusted to make it look closer to how the font icon
renders in VS Code. Unfortunately the spacing is not perfect, and this
now also adds a useless entry to the "Select Language" selector.

It would be great if https://github.com/jesseweed/seti-ui/pull/716 could
get merged as that would fix all these problems.